+++ src/components/Accounts/Actions/PayModal.js
@@ -1,6 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
+import {
+  FormattedMessage,
+  injectIntl,
+  intlShape,
+} from 'react-intl';
 import { Field, reduxForm, change } from 'redux-form';
 import {
   Row,
@@ -22,19 +26,19 @@ const validate = (values, props) => {
 
   const errors = {};
   if (!values.amount) {
-    errors.amount = props.stripes.intl.formatMessage({ id: 'ui-users.accounts.error.field' });
+    errors.amount = <FormattedMessage id="ui-users.accounts.error.field" />;
   }
   if (values.amount < 0) {
-    errors.amount = props.stripes.intl.formatMessage({ id: 'ui-users.accounts.pay.error.amount' });
+    errors.amount = <FormattedMessage id="ui-users.accounts.pay.error.amount" />;
   }
   if (!values.method) {
     errors.method = 'Select one';
   }
   if (props.commentRequired && !values.comment) {
-    errors.comment = props.stripes.intl.formatMessage({ id: 'ui-users.accounts.error.comment' });
+    errors.comment = <FormattedMessage id="ui-users.accounts.error.comment" />;
   }
   if (values.amount > selected) {
-    errors.amount = props.stripes.intl.formatMessage({ id: 'ui-users.accounts.error.exceeds' });
+    errors.amount = <FormattedMessage id="ui-users.accounts.error.exceeds" />;
   }
   return errors;
 };
@@ -54,9 +58,7 @@ class PayModal extends React.Component {
     reset: PropTypes.func,
     dispatch: PropTypes.func,
     commentRequired: PropTypes.bool,
-    stripes: PropTypes.shape({
-      intl: PropTypes.object.isRequired,
-    }),
+    intl: intlShape.isRequired,
   };
 
   constructor(props) {
@@ -105,11 +107,17 @@ class PayModal extends React.Component {
     parseFloat(selected).toFixed(2);
     const remaining = parseFloat(totalamount - this.state.amount).toFixed(2);
     const payments = this.props.payments.map(p => ({ id: p.id, label: p.nameMethod }));
-    const { submitting, invalid, pristine } = this.props;
+    const {
+      submitting,
+      invalid,
+      pristine,
+      intl,
+    } = this.props;
     const paymentAmount = this.state.amount === '' ? 0.00 : this.state.amount;
     const message = `${(this.state.amount < selected) ? 'Partially paying' : 'Paying'} ${n} ${(n === 1) ? 'fee/fine' : 'fees/fines'} for a total amount of ${parseFloat(paymentAmount).toFixed(2)}`;
-    const additional = this.props.stripes.intl.formatMessage({ id: 'ui-users.accounts.pay.placeholder.additional' });
-    const comment = `${additional} ${(this.props.commentRequired) ? '(required)' : '(optional)'}`;
+    const comment = (this.props.commentRequired)
+      ? intl.formatMessage({ id: 'ui-users.accounts.pay.placeholder.additional.required' })
+      : intl.formatMessage({ id: 'ui-users.accounts.pay.placeholder.additional.optional' });
 
     return (
       <Modal
@@ -229,4 +237,4 @@ export default reduxForm({
   form: 'payment',
   fields: [],
   validate,
-})(PayModal);
+})(injectIntl(PayModal));